Creacion de Nuevas clases
Se creo la clase Ant Historial
Se creo la clase General historial

+++ AntProject/src/EDD/Path.java
@@ -55,7 +55,7 @@ public class Path {
         return pheromones;
     }
 
-    public void setPheromones(int pheromones) {
+    public void setPheromones(double pheromones) {
         this.pheromones = pheromones;
     }
